# server/docs/swagger.yaml
openapi: 3.0.1
info:
  title: XenoReach CRM API
  version: 1.0.0
  description: |
    API documentation for XenoReach CRM. Use the Authorize button to enter your JWT Bearer token for protected endpoints.
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Segment:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        rules: { type: object }
        mongoQuery: { type: object, nullable: true }
        customerCount: { type: integer }
        isActive: { type: boolean }
        lastRun: { type: string, format: date-time, nullable: true }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
paths:
...
  /api/segments:
    get:
      summary: List all segments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of segments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Segment'
    post:
      summary: Create a new segment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                rules: { type: object }
      responses:
        '201':
          description: Created
  /api/segments/preview:
    post:
      summary: Preview segment (get count/sample)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rules: { type: object }
      responses:
        '200':
          description: Preview result
...
  /api/analytics/segments:
    get:
      summary: Get segment analytics
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Segment analytics
...